// apps/web-panel-vue/src/views/site/types.d.ts
export interface RiskTableDataProps {
	id: number
	content: string
	new_content_file: string
	risk_content: string
	risk_type: string
	site_id: number
	site_name: string
	source_content_file: string
	source_file: string
	testing_id: string
	url: string
	time: number
}